Craft system, autoCraft, optimization, containers upgrade

+++ Assets/SkyArchipelago/Scripts/Common/Signals/Containers/MoveItemBetweenContainersSignal.cs
@@ -2,9 +2,9 @@ public struct MoveItemBetweenContainersSignal
 {
     public int ContainerIdFrom;
     public int ContainerIdTo;
-    public byte FromIdSlot;
+    public int FromIdSlot;
 
-    public MoveItemBetweenContainersSignal(int from, int to, byte slotId)
+    public MoveItemBetweenContainersSignal(int from, int to, int slotId)
     {
         ContainerIdFrom = from;
         ContainerIdTo = to;
